+++ CHANGELOG.md
@@ -11,10 +11,14 @@ Non-collectable elements are various sub-elements to collectable elements.
 
 #### XML - Constant elements
 
+* ApplicationRuleBasedValueSpecification | APPLICATION-RULE-BASED-VALUE-SPECIFICATION
+* CompositeRuleBasedValueSpecification | COMPOSITE-RULE-BASED-VALUE-SPECIFICATION
 * NumericalOrText | NUMERICAL-OR-TEXT
 * NumericalRuleBasedValueSpecification | NUMERICAL-RULE-BASED-VALUE-SPECIFICATION
 * ReferenceValueSpecification | REFERENCE-VALUE-SPECIFICATION
 * RuleArguments | RULE-ARGUMENTS
+* RuleBasedAxisCont | RULE-BASED-AXIS-CONT
+* RuleBasedValueCont | RULE-BASED-VALUE-CONT
 * RuleBasedValueSpecification | RULE-BASED-VALUE-SPECIFICATION
 
 #### Developer Utilities (`dev_utils`)